requirements.txt
asgiref==3.5.0
async-generator==1.10
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
cryptography==36.0.1
decouple==0.0.7
Django==4.0.2
django-tinymce==3.4.0
h11==0.13.0
idna==3.3
outcome==1.1.0
psycopg2-binary==2.9.3
pycparser==2.21
pyOpenSSL==22.0.0
selenium==4.1.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.8
wsproto==1.0.0
